Flatten Unmapped partial class to a single file
